backend/requirements.txt
flask
numpy 
tensorflow-macos
tensorflow-metal 
opencv-python
Pillow